refactor(i18n): move the catalog generator out of src to codegen/i18n
The generator is Node-only build tooling that reads the filesystem -- the one thing this
SDK's source must never do, since a Node-only import breaks browser and React Native
bundles outright. It sat in `src/` as a sibling of `src/i18n/` with only a build-time regex
keeping the runtime layer away from it. It is now at `codegen/i18n/`, outside `src/`
entirely.

It stays published: two other repos import `stream-chat/i18n/codegen` from their build
scripts, so this is versioned, typed, semver-relevant API rather than internal tooling like
`scripts/bundle.mjs`. **The published surface is unchanged** -- `dist/types/` is
byte-identical, the root and i18n bundles are byte-identical (907,599), and `exports` /
`typesVersions` still point at `dist/types/i18n-codegen/`. The only diff in `dist/` is the
source paths esbuild embeds in the two codegen bundles, which got shorter.

The move is cheap because the generator was already fully decoupled: zero imports from
`src/`, only `node:fs`, `node:path`, a type-only `typescript` and its own siblings. Nothing
in `src/` imported it either.

Three things were scoped to `src/` and would each have failed **silently**:

- `tsconfig.json` has `rootDir: "./src"` and `include: ["./src/**/*"]`, which is what maps the
  generator's declarations to `dist/types/i18n-codegen/`. Widening `include` would not work --
  `rootDir` must contain every input, so it would become `.` and every path under
  `dist/types` would gain a `src/` prefix, breaking all the types entries. Hence
  `tsconfig.codegen.json`, with `rootDir: "./codegen/i18n"` so the output path is unchanged.
- `yarn types` was bare `tsc --noEmit` on the root project; the generator would have stopped
  being typechecked. Both `types` and `build` now run both projects.
- Every `eslint.config.mjs` rule block is `files: ['src/**/*.{js,ts}']`, and the top-level
  `ignores` has `'*.{js,ts}'`, so a new top-level directory inherits no rules whatsoever.
  Both blocks now list `codegen/**/*.{js,ts}`.

Verified rather than assumed: a `!` in `codegen/i18n/` is now reported by `yarn eslint`, a type
error there is now reported by `yarn types`, and the packed tarball's
`stream-chat/i18n/codegen` resolves with all ten exports and its `.d.ts` present.

One real gain beyond tidiness: the "i18n runtime must not reach the generator" boundary is now
enforced by the type system, since `codegen/` is outside the library project -- an import from
`src/i18n/` fails at `tsc` before the metafile assertion runs. The error is an oblique TS6059
"not under rootDir", so the assertion in `scripts/bundle.mjs` stays as the backstop, and its
comment now says why.
